Draw the one link a delete file records, and let last-sequence-number be long
A v3 deletion vector carries referenced_data_file, which names exactly one data
file. That is the only delete-to-data link in the format — a positional delete
file's targets are inside its rows, one per row, and an equality delete has none
at all, it applies by predicate — so it is the only one the graph can honestly
draw. The file inspector has been explaining this for two commits; now the edge
is there.

affectsLayout = false, for the reason commit lineage is: both ends are data
files and belong in the same layer, and letting ELK see the edge would push the
referenced file a layer right of a node it sits beside. Rendered and looked at:
two red curves inside the file column, from each .puffin delete to the parquet
file it names.

One data file can have a node under more than one manifest, when several
retained manifests still list it. Each of those nodes is that file, so each gets
the edge. Drawing one and choosing which would be a claim about which manifest
the vector belongs to, and the vector records no such thing.

Separately, TableMetadata.lastSequenceNumber was Int where the spec says long,
which is not a misread but a refusal — one out-of-range field and the whole
document fails to parse, taking the table with it. Every other sequence number
here is already Long.

The canvas render now writes two files. graph-canvas-partial is a graph the page
size has cut down, which is the state the badge exists for; graph-canvas-whole
is one drawn entire, which is the only way both ends of a deletion-vector edge
are on screen.

+++ core/src/main/kotlin/model/IcebergSchema.kt
@@ -12,7 +12,9 @@ data class TableMetadata(
     @SerialName("format-version") val formatVersion: Int? = null,
     @SerialName("table-uuid") val tableUuid: String? = null,
     val location: String? = null,
-    @SerialName("last-sequence-number") val lastSequenceNumber: Int? = null,
+    // long in the spec, and every other sequence number here is a Long. Unreachable in practice
+    // — it would take 2^31 commits — but a table that got there would fail to parse at all.
+    @SerialName("last-sequence-number") val lastSequenceNumber: Long? = null,
     @SerialName("last-updated-ms") val lastUpdatedMs: Long? = null,
     @SerialName("last-column-id") val lastColumnId: Int? = null,
     @SerialName("current-schema-id") val currentSchemaId: Int? = null,
